+++ trivia-mvc/Models/TriviasQuestionsReceived.cs
@@ -7,8 +7,8 @@ namespace trivia_mvc.Models
 {
     public partial class TriviasQuestionsReceived
     {
-        public short IdTrivia { get; set; }
-        public short IdQuestion { get; set; }
+        public int IdTrivia { get; set; }
+        public int IdQuestion { get; set; }
 
         public virtual Question IdQuestionNavigation { get; set; }
         public virtual Trivia IdTriviaNavigation { get; set; }
